requirements.txt
backports.zoneinfo==0.2.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
dateparser==1.2.0
filelock==3.16.1
Flask==3.0.3
fsspec==2024.9.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-assistant-grpc==0.0.2
google-assistant-sdk==0.3.3
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.1.0
google-cloud-speech==2.27.0
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.41.1
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
openai-whisper @ git+https://github.com/openai/whisper.git@25639fc17ddc013d56c594bfbf7644f2185fad84
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.10.1
six==1.16.0
sounddevice==0.3.7
SpeechRecognition==3.10.4
sympy==1.13.3
tenacity==4.1.0
tiktoken==0.7.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.5
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.21.1
Werkzeug==3.0.4
zipp==3.20.2
